% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdrreg_output.r
\name{print.fdrreg_result}
\alias{print.fdrreg_result}
\title{Print an fdrreg_result Object}
\usage{
\method{print}{fdrreg_result}(x, ...)
}
\arguments{
\item{x}{An object of class \code{fdrreg_result}.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print an fdrreg_result Object
}
