% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdrreg_output.r
\name{summary.fdrreg_result}
\alias{summary.fdrreg_result}
\title{Summarize an fdrreg_result Object}
\usage{
\method{summary}{fdrreg_result}(object, ...)
}
\arguments{
\item{object}{An object of class \code{fdrreg_result}.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns \code{object}.
}
\description{
Summarize an fdrreg_result Object
}
