% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.r
\name{load_builtin_annotations}
\alias{load_builtin_annotations}
\title{Load Built-in Curated Psychiatric Annotations}
\usage{
load_builtin_annotations(key_type = c("entrez", "ensembl", "symbol"))
}
\arguments{
\item{key_type}{Character, identifier column to use: "entrez" (ID), "ensembl" (ENSEMBL_GENE_ID), or "symbol" (GeneName).}
}
\value{
A data frame containing gene identifiers and 21 biological annotation columns.
}
\description{
Retrieves the frozen annotation matrix provided in Supplementary Table S1.21.
}
\examples{
annot <- load_builtin_annotations("entrez")
head(annot[, 1:5])
}
